# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Epic Poetry Cafe Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expenses
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: fromDate
          in: query
          schema:
            type: string
        - name: toDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    get:
      operationId: getExpense
      tags: [expenses]
      summary: Get expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
...
  /petty-cash:
    get:
      operationId: listPettyCash
      tags: [pettyCash]
      summary: List petty cash transactions
      parameters:
        - name: fromDate
          in: query
          schema:
            type: string
        - name: toDate
          in: query
          schema:
            type: string
        - name: transactionType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PettyCashTransaction"
    post:
      operationId: createPettyCash
      tags: [pettyCash]
      summary: Create petty cash transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePettyCashRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PettyCashTransaction"
  /petty-cash/{id}:
    delete:
      operationId: deletePettyCash
      tags: [pettyCash]
      summary: Delete petty cash transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /petty-cash/summary:
    get:
      operationId: getPettyCashSummary
      tags: [pettyCash]
      summary: Petty cash balance summary
      parameters:
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PettyCashSummary"
components:
  schemas:
...
    Expense:
      type: object
      properties:
        id:
          type: integer
        expenseNumber:
          type: string
        expenseDate:
          type: string
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        vendorId:
          type: ["integer", "null"]
        vendorName:
          type: ["string", "null"]
        amount:
          type: number
        taxAmount:
          type: number
        totalAmount:
          type: number
        paymentMode:
          type: ["string", "null"]
        paidBy:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        costType:
          type: string
        recurring:
          type: boolean
        recurringFrequency:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - expenseNumber
        - expenseDate
        - amount
        - totalAmount
        - costType
    CreateExpenseRequest:
      type: object
      properties:
        expenseDate:
          type: string
        categoryId:
          type: integer
        vendorId:
          type: integer
        amount:
          type: number
        taxAmount:
          type: number
        paymentMode:
          type: string
        paidBy:
          type: string
        description:
          type: string
        costType:
          type: string
        recurring:
          type: boolean
        recurringFrequency:
          type: string
      required:
        - expenseDate
        - amount
        - costType
...
    PettyCashTransaction:
      type: object
      properties:
        id:
          type: integer
        transactionDate:
          type: string
        transactionType:
          type: string
        amount:
          type: number
        method:
          type: ["string", "null"]
        counterpartyName:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        linkedExpenseId:
          type: ["integer", "null"]
        description:
          type: ["string", "null"]
        runningBalance:
          type: number
        approvalStatus:
          type: string
        createdBy:
          type: ["integer", "null"]
        approvedBy:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - transactionDate
        - transactionType
        - amount
        - runningBalance
        - approvalStatus
    CreatePettyCashRequest:
      type: object
      properties:
        transactionDate:
          type: string
        transactionType:
          type: string
          enum: [receipt, expense, adjustment]
        amount:
          type: number
        method:
          type: string
        counterpartyName:
          type: string
        category:
          type: string
        linkedExpenseId:
          type: integer
        description:
          type: string
      required:
        - transactionDate
        - transactionType
        - amount
    PettyCashSummary:
      type: object
      properties:
        openingBalance:
          type: number
        totalReceipts:
          type: number
        totalExpenses:
          type: number
        totalAdjustments:
          type: number
        closingBalance:
          type: number
        currentBalance:
          type: number
        transactionCount:
          type: integer
      required:
        - openingBalance
        - totalReceipts
        - totalExpenses
        - totalAdjustments
        - closingBalance
        - currentBalance
        - transactionCount