# lib/api-zod/src/generated/types/createPettyCashRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Epic Poetry Cafe Management API
 * OpenAPI spec version: 0.1.0
 */
import type { CreatePettyCashRequestTransactionType } from "./createPettyCashRequestTransactionType";

export interface CreatePettyCashRequest {
  transactionDate: string;
  transactionType: CreatePettyCashRequestTransactionType;
  amount: number;
  method?: string;
  counterpartyName?: string;
  category?: string;
  linkedExpenseId?: number;
  description?: string;
}