# lib/api-zod/src/generated/types/pettyCashTransaction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Epic Poetry Cafe Management API
 * OpenAPI spec version: 0.1.0
 */

export interface PettyCashTransaction {
  id: number;
  transactionDate: string;
  transactionType: string;
  amount: number;
  /** @nullable */
  method?: string | null;
  /** @nullable */
  counterpartyName?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  linkedExpenseId?: number | null;
  /** @nullable */
  description?: string | null;
  runningBalance: number;
  approvalStatus: string;
  /** @nullable */
  createdBy?: number | null;
  /** @nullable */
  approvedBy?: number | null;
  createdAt?: Date;
}